// packages/ketcher-react/src/script/ui/views/toolbars/toolbar.types.ts
/****************************************************************************
 * Copyright 2021 EPAM Systems
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *    http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 ***************************************************************************/
import type { MULTITAIL_ARROW_TOOL_NAME } from 'ketcher-core';

type TopGroup = 'document' | 'edit' | 'zoom' | 'process' | 'meta';

type LeftGroup =
  | 'hand'
  | 'select'
  | 'bond'
  | 'charge'
  | 'transform'
  | 'sgroup'
  | 'rgroup';

type BottomGroup = 'template-common' | 'template-lib';

type RightGroup = 'atom' | 'period-table';

type ToolbarGroupVariant = TopGroup | LeftGroup | BottomGroup | RightGroup;

type TopToolbarItemVariant =
  | 'clear'
  | 'open'
  | 'save'
  | 'undo'
  | 'redo'
  | 'cut'
  | 'copies'
  | 'copy'
  | 'copy-image'
  | 'copy-mol'
  | 'copy-ket'
  | 'paste'
  | 'zoom-in'
  | 'zoom-out'
  | 'zoom-list'
  | 'layout'
  | 'clean'
  | 'arom'
  | 'dearom'
  | 'cip'
  | 'check'
  | 'help'
  | 'about';

type LeftToolbarItemVariant =
  // select group
  | 'hand'
  | 'select'
  | 'select-lasso'
  | 'select-rectangle'
  | 'select-structure'
  | 'select-fragment'
  | 'erase'
  // bond group
  | 'bonds'
  | 'bond-common'
  | 'bond-single'
  | 'bond-double'
  | 'bond-triple'
  | 'bond-dative'
  | 'bond-stereo'
  | 'bond-up'
  | 'bond-down'
  | 'bond-updown'
  | 'bond-crossed'
  | 'bond-query'
  | 'bond-special'
  | 'bond-any'
  | 'bond-hydrogen'
  | 'bond-aromatic'
  | 'bond-singledouble'
  | 'bond-singlearomatic'
  | 'bond-doublearomatic'
  | 'chain'
  // charge group
  | 'charge-plus'
  | 'charge-minus'
  // radical group
  | 'radical'
  // sgroup group
  | 'sgroup'
  // reaction
  // arrows
  | 'arrows'
  | typeof MULTITAIL_ARROW_TOOL_NAME
  // mapping
  | 'reaction-mapping-tools'
  | 'reaction-automap'
  // rgroup group
  | 'rgroup'
  | 'rgroup-attpoints'
  // create monomer
  | 'create-monomer';

type BottomToolbarItemVariant =
  | 'template-common'
  | 'template-lib'
  | 'enhanced-stereo';

type RightToolbarItemVariant =
  | 'atom'
  | 'freq-atoms'
  | 'period-table'
  | 'extended-table'
  | 'any-atom';

type FloatingToolItemVariant =
  | 'transform-flip-h'
  | 'transform-flip-v'
  | 'erase';

type ToolbarItemVariant =
  | TopToolbarItemVariant
  | LeftToolbarItemVariant
  | BottomToolbarItemVariant
  | RightToolbarItemVariant
  | FloatingToolItemVariant
  | 'bonds';

interface ToolbarItem {
  id: ToolbarItemVariant;
  options?: ToolbarItem[];
}

export type { ToolbarGroupVariant };

export type {
  BottomToolbarItemVariant,
  LeftToolbarItemVariant,
  RightToolbarItemVariant,
  TopToolbarItemVariant,
  FloatingToolItemVariant,
  ToolbarItemVariant,
};

export type { ToolbarItem };